DjangoRESTFirst/requirements.txt
django
djangorestframework
django-filter
django-guardian
markdown
pygments
mysqlclient